HTTP: refactor: storing the body_handler as part of r.
This will allow sending the header from a totally different point, since
the data for the call is present in the request, which is available
everywhere.

It will also allow consulting in a filter if there is a body_handler
installed.  The gzip filter will need this, as it should be a no-op if
there is no body handler installed.

Signed-off-by: Alejandro Colomar <alx@nginx.com>