Wasm: using offsetof(3) to compute the offset of the flex array.
Using sizeof() may give incorrect results if there's trailing padding
in the structure and it overlaps with the flexible array member.

It worked before this patch, because there was no padding, but that's a
brittle thing to rely on.

While preparing an equivalent patch for the Linux kernel --which also
used sizeof() for its implementation of struct_size()--, we immediately
found one bug (see the LKML link below), due to offsetof(3) being more
restrictive in its input than sizeof().

`nxt_max(sizeof(s), ...)` is necessary to avoid allocating less than the
structure size, if the flex array is too small.

Link: <https://gustedt.wordpress.com/2011/03/14/flexible-array-member/>
Link: <https://lore.kernel.org/lkml/202308161913.91369D4A@keescook/T/>
Link: <https://inbox.sourceware.org/gcc/dac8afb7-5026-c702-85d2-c3ad977d9a48@kernel.org/T/>
Link: <https://github.com/nginx/unit/pull/922>
Cc: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>