Using offsetof(3) to compute the offset of the flex array member.
Use the nxt_sizeof_struct() wrapper --and related ones-- where
appropriate, instead of open-coding their patterns.

Using sizeof() may give incorrect results if there's trailing padding
in the structure and it overlaps with the flexible array member.

It worked before this patch, because there was no padding, but that's a
brittle thing to rely on, as new structure members may change the
padding.  Or even if some structure is stable, using the same pattern
calculation in a different structure may not be similarly safe.  It
seems safer to me to always use offsetof(3) --and through the recently-
added wrappers where appropriate-- to be on the safe side.

While preparing an equivalent patch for the Linux kernel --the kernel
also used sizeof() for its implementation of struct_size(), so the patch
reimplemented it in terms of offsetof(3)--, we immediately found one bug
(see the LKML link below), due to offsetof(3) being more restrictive in
its input than sizeof().  So this does find real bugs.

Link: <https://gustedt.wordpress.com/2011/03/14/flexible-array-member/>
Link: <https://lore.kernel.org/lkml/202308161913.91369D4A@keescook/T/>
Link: <https://inbox.sourceware.org/gcc/dac8afb7-5026-c702-85d2-c3ad977d9a48@kernel.org/T/>
Link: <https://github.com/nginx/unit/pull/922>
Cc: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>