Mp: using _Alignof(max_align_t) for aligning a size.
sizeof() is not appropriate for alignment.  _Alignof() is.

Regarding the type used for alignment, 'max_align_t' was designed for
this specific use: get the largest alignment of a basic type.  That
'uintptr_t' happens to have the same alignment is just a coincidence,
but it's not appropriate to use it.

Signed-off-by: Alejandro Colomar <alx@nginx.com>