Libunit: fixed size calculation.
This calculation intended to get the position at the end of the
'nxt_unit_response_t' structure, after the flexible array member, where
the strings for the names and values of the header fields are stored.

However, it was missing an offsetof(), and was getting a pointer to
somewhere else (earlier in memory).  Thus, somewhere in the middle of
the structure was being clobbered.

Use nxt_sizeof_struct(), which performs the right offsetof() calculation
internally.

Closes: <https://github.com/nginx/unit/issues/923>
Fixes: 1bb22d1e922c ("Unit application library.")
Signed-off-by: Alejandro Colomar <alx@nginx.com>