Added routes array to the default configuration.
The default configuration previously contained just a listeners and
applications object. Since routes is now a principle configuration object,
and a recommended way of configurating Unit, it is now included in the
default configuration.

This change benefits new users because it explicitly introduces the three
principle configuration objects which leads more intuitively to the
documentation. Experienced users may choose to ignore or delete routes.

routes is defined as an array instead of an object because this change
is designed to assist new users, where the simpler form of routes is
easier to understand.