PHP: Fix a possible file-pointer leak.
In nxt_php_execute() it is possible we could bail out before cleaning up
the FILE * representing the PHP script to execute.

At this point we only need to call fclose(3) on it.

We could have possibly moved the opening of this file to later in the
function, but it is probably good to bail out as early as possible if we
can't open it.

This was found by Coverity.

Fixes: bebc03c72 ("PHP: Implement better error handling.")
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>