.mailmap: Set correct address for Danielle
Due to GH making a mess of merge commits, it used Danielle's personal
email address for the merge, it also used a generic GH address for the
committer but we can't do anything about that. However we can fix the
'Author' email address.

If for some reason you want to see the original names/addresses used you
can generally pass --no-mailmap to git commands.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>