Docker: Add a new target for generic language versions
You can use the Docker Makefile to create dockerfiles for specific
language environments, e.g

  $ make build-python3.12 VERSIONS_python=3.12

However this breaks when trying to do something like

  $ make build-php8.3.0RC6 VERSIONS_php=8.3.0RC6 VARIANT_php=cli

This breaks due to the RC in the version, due to how we're trying to
parse out the module name from the version string, we strip out anything
that is a 0-9 or a . or -.

That works for 'python3.12' leaving 'python'. However for 'php8.3.0RC6'
that leaves 'phpRC'.

After spending some time on this, the best (and simplest) solution I
could come up with, that doesn't break existing usage, is to create a
new make target for this.

So the above now becomes

  $ make gen-php8.3.0RC6 MODULE=php VERSIONS_php=8.3.0RC6 VARIANT_php=cli

The new target is gen-<language version> and you need to specify the
language module.

Closes: https://github.com/nginx/unit/issues/1009
Tested-by: Timo Stark <t.stark@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>