Static: Don't include the terminating null byte in exten
The 'path' is (or should be) terminated by a null byte ('\0').  But the
code reading 'exten' assumes there's no null byte.  Thus, we must reduce
the length of that string by 1.

The code specialized for 'index' (which happens to be the only code that
didn't have a bug) must now use the 'path' variable, which includes the
null byte.

The code for non-index cases, which for other reasons assumes 'shr' is
null-terminated, in this call assumes it's not null terminated.  Yes,
it's quite buggy in that regard.  That was a bug introduced in
c5220944d2ac ("Static: variables in the "share" option.").

That commit included some important refactor alongside a the feature
that it adds.  That resulted in the bug we're (partially) fixing now.

    $ git show c5220944d2ac -- src/nxt_http_static.c | grep exten;
    -    nxt_str_t               index, exten, *mtype;
    +    nxt_str_t               *shr, exten, *mtype;
    -        nxt_str_set(&exten, ".html");
    -        nxt_str_null(&exten);
    -    if (conf->types != NULL && exten.start == NULL) {
    -        nxt_http_static_extract_extension(r->path, &exten);
    -        mtype = nxt_http_static_mtype_get(&rtcf->mtypes_hash, &exten);
    +        nxt_str_set(&exten, ".html");
    +            nxt_str_null(&exten);
    +            nxt_http_static_extract_extension(shr, &exten);
    +            mtype = nxt_http_static_mtype_get(&rtcf->mtypes_hash, &exten);
             if (exten.start == NULL) {
    -            nxt_http_static_extract_extension(r->path, &exten);
    +            nxt_http_static_extract_extension(shr, &exten);

Above we can see that calls to this function changed the input from
r->path (which is non-terminated) to shr (which should be), without
changing the implementation of this function.

    $ git show c5220944d2ac -- src/nxt_http_static.c | grep exten;
    -    nxt_str_t               index, exten, *mtype;
    +    nxt_str_t               *shr, exten, *mtype;
    -        nxt_str_set(&exten, ".html");
    -        nxt_str_null(&exten);
    -    if (conf->types != NULL && exten.start == NULL) {
    -        nxt_http_static_extract_extension(r->path, &exten);
    -        mtype = nxt_http_static_mtype_get(&rtcf->mtypes_hash, &exten);
    +        nxt_str_set(&exten, ".html");
    +            nxt_str_null(&exten);
    +            nxt_http_static_extract_extension(shr, &exten);
    +            mtype = nxt_http_static_mtype_get(&rtcf->mtypes_hash, &exten);
             if (exten.start == NULL) {
    -            nxt_http_static_extract_extension(r->path, &exten);
    +            nxt_http_static_extract_extension(shr, &exten);

Above we can see that fname, which prior to c5220944d2ac was always a
null-terminated string (it is later passed to open(2)), was originally
formed as `shr [+ index] + '\0'`.  After c5220944d2ac, it changed to be
just `shr` (under some conditions), which is not terminated.

In this commit, we're partially fixing that bug.  We're adapting this
function to handle a null-terminated string.

However, there's still a remaining bug: some inputs to this function are
incorrectly not terminated, and they must be terminated by a null byte
(because we're also passing them to open(2)).  That should be fixed in a
future commit.

Fixes: c5220944d2ac ("Static: variables in the "share" option.")
Cc: Andrew Clayton <a.clayton@nginx.com>
Cc: Zhidao Hong <z.hong@f5.com>
Cc: Andrei Zeliankou <zelenkov@nginx.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>