Static: Terminate string with a null byte
This string is later passed to open(2).  It is also passed to
nxt_http_static_extract_extension(), which also requires it to be
terminated.

This bug was introduced in c5220944d2ac, but for some reason was not
discovered until recently, when I triggered it purposefully two commits
ago, while partially fixing the bug (or actually fixing another related
bug, which was introduced in the same commit).

Fixes: c5220944d2ac ("Static: variables in the "share" option.")
Fixes: Two commits ago ("Static: Don't include the terminating null byte in exten")
Cc: Andrew Clayton <a.clayton@nginx.com>
Cc: Zhidao Hong <z.hong@f5.com>
Cc: Andrei Zeliankou <zelenkov@nginx.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>