Tstr: Produce a nxt_strz_t from nxt_tstr_t
In the previous commit, we made sure all nxt_tstr_t have a hidden NUL at
the end of the produced nxt_str_t (hidden because the .length field
doesn't include it).

We can make that a different type, nxt_strz_t, which is distinct from
nxt_str_t.  This will allow the compiler to enforce correct use of these
strings, via type-safety diagnostics.

nxt_strz_t can be read as a nxt_str_t, for compatibility in functions
that want a nxt_str_t.  This access is through a const member of a union
to ensure that the terminated string is not corrupted.

Cc: Andrew Clayton <a.clayton@nginx.com>
Cc: Zhidao Hong <z.hong@f5.com>
Cc: Andrei Zeliankou <zelenkov@nginx.com>
Cc: "Valentin V. Bartenev" <vbartenev@gmail.com>
Cc: Dan Callahan <d.callahan@f5.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>