Isolation: Add a new nxt_cred_t type
This is a generic type to represent a uid_t/gid_t on Linux when user
namespaces are in use.

Technically this only needs to be an unsigned int, but we make it an
int64_t so we can make use of the existing NXT_CONF_MAP_INT64 type.

This will be used in subsequent commits.

Reviewed-by: Zhidao Hong <z.hong@f5.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>